nutrisnap/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-telegram-bot==21.9
google-generativeai>=0.8.0
langgraph>=0.2.0
langchain-core>=0.3.0
httpx==0.27.0
pydantic>=2.0.0
python-dotenv==1.0.0
Pillow>=10.0.0
pyzbar==0.1.9
